Moved plot function up to parent class.
